requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-groq
langchain-text-splitters
chromadb
sentence-transformers
pypdf
pandas
python-dotenv